feat: rewrite hotbar to React InventoryWindow component
