+++ src/react/HotbarRenderApp.tsx
@@ -1,15 +1,23 @@
-import { useEffect, useRef, useState } from 'react'
+import { useEffect, useMemo, useState } from 'react'
 import { motion, AnimatePresence } from 'framer-motion'
 import { createPortal } from 'react-dom'
-import { subscribe, useSnapshot } from 'valtio'
-import { activeModalStack, isGameActive, miscUiState } from '../globalState'
-import { currentScaling } from '../scaleInterface'
-import { watchUnloadForCleanup } from '../gameUnload'
+import { useSnapshot } from 'valtio'
+import {
+  TextureProvider,
+  ScaleProvider,
+  InventoryProvider,
+  InventoryWindow,
+  createMineflayerConnector,
+  type MineflayerBot,
+} from 'minecraft-inventory/src'
+import { activeModalStack, miscUiState } from '../globalState'
+import { useAppScale } from '../scaleInterface'
 import { getItemNameRaw } from '../mineflayer/items'
 import { isInRealGameSession } from '../utils'
 import { triggerCommand } from '../controls'
 import MessageFormattedString from './MessageFormattedString'
 import SharedHudVars from './SharedHudVars'
+import { textureConfig, buildItemMapper, clearInventoryCaches } from './inventory/sharedConnectorSetup'
 
 export const BASE_HOTBAR_HEIGHT = 25
 
@@ -72,79 +80,25 @@ const ItemName = ({ itemKey }: { itemKey: string }) => {
 }
 
 const HotbarInner = () => {
-  const container = useRef<HTMLDivElement>(null!)
   const [itemKey, setItemKey] = useState('')
+  const [textureVersion, setTextureVersion] = useState(0)
   const hasModals = useSnapshot(activeModalStack).length
   const { currentTouch, appConfig } = useSnapshot(miscUiState)
-  const mobileOpenInventory = currentTouch && !appConfig?.disabledCommands?.includes('general.inventory')
+  const appScale = useAppScale()
+
+  const supportsOffhand = !bot.supportFeature('doesntHaveOffHandSlot')
+  const isMobile = currentTouch && !appConfig?.disabledCommands?.includes('general.inventory')
+
+  const connector = useMemo(() => {
+    return createMineflayerConnector(bot as MineflayerBot, {
+      itemMapper: buildItemMapper(bot.version),
+    })
+  }, [textureVersion])
 
   useEffect(() => {
     const controller = new AbortController()
 
-    // @ts-expect-error -- openItemsCanvas removed during inventory rework (PR #518). HotbarInner is dead code (returns null).
-    const inv = openItemsCanvas('HotbarWin', {
-      _client: {
-        write () {}
-      },
-      clickWindow (slot, mouseButton, mode) {
-        if (mouseButton === 1) {
-          console.log('right click')
-          return
-        }
-        const hotbarSlot = slot - bot.inventory.hotbarStart
-        if (hotbarSlot < 0 || hotbarSlot > 8) return
-        bot.setQuickBarSlot(hotbarSlot)
-      },
-    } as any)
-    const { canvasManager } = inv
-    inv.inventory.supportsOffhand = !bot.supportFeature('doesntHaveOffHandSlot')
-    inv.pwindow.disablePicking = true
-
-    canvasManager.children[0].disableHighlight = true
-    canvasManager.minimizedWindow = true
-    canvasManager.minimizedWindow = true
-
-    function setSize () {
-      canvasManager.setScale(currentScaling.scale)
-
-      canvasManager.windowHeight = BASE_HOTBAR_HEIGHT * canvasManager.scale
-      canvasManager.windowWidth = (210 - (inv.inventory.supportsOffhand ? 0 : 25) + (mobileOpenInventory ? 28 : 0)) * canvasManager.scale
-    }
-    setSize()
-    watchUnloadForCleanup(subscribe(currentScaling, setSize))
-    inv.canvas.style.pointerEvents = 'auto'
-    container.current.appendChild(inv.canvas)
-    const upHotbarItems = () => {
-      if (!appViewer.resourcesManager?.itemsAtlasParser) return
-      // @ts-expect-error -- upInventoryItems removed during inventory rework (PR #518). HotbarInner is dead code (returns null).
-      globalThis.debugHotbarItems = upInventoryItems(true, inv)
-    }
-
-    canvasManager.canvas.onclick = (e) => {
-      if (!isGameActive(true)) return
-      const pos = inv.canvasManager.getMousePos(inv.canvas, e)
-      if (canvasManager.canvas.width - pos.x < 35 * inv.canvasManager.scale && mobileOpenInventory) {
-        triggerCommand('general.inventory', true)
-        triggerCommand('general.inventory', false)
-      }
-    }
-
-    globalThis.debugUpHotbarItems = upHotbarItems
-    upHotbarItems()
-    bot.inventory.on('updateSlot', upHotbarItems)
-    appViewer.resourcesManager.on('assetsTexturesUpdated', upHotbarItems)
-    appViewer.resourcesManager.on('assetsInventoryReady', () => {
-      upHotbarItems()
-    })
-
-    const setSelectedSlot = (index: number) => {
-      if (index === bot.quickBarSlot) return
-      bot.setQuickBarSlot(index)
-      if (!bot.inventory.slots?.[bot.quickBarSlot + 36]) setItemKey('')
-    }
     const heldItemChanged = () => {
-      inv.inventory.activeHotbarSlot = bot.quickBarSlot
-
       if (!bot.inventory.slots?.[bot.quickBarSlot + 36]) {
         setItemKey('')
         return
@@ -153,49 +107,54 @@ const HotbarInner = () => {
       const itemNbt = item.nbt ? JSON.stringify(item.nbt) : ''
       setItemKey(`${item.name}_split_${item.type}_split_${item.metadata}_split_${itemNbt}_split_${JSON.stringify(item['components'] ?? [])}`)
     }
-    heldItemChanged()
+    heldItemChanged() // initial call
     bot.on('heldItemChanged' as any, heldItemChanged)
 
     document.addEventListener('wheel', (e) => {
       if (!isInRealGameSession()) return
       e.preventDefault()
       const newSlot = ((bot.quickBarSlot + Math.sign(e.deltaY)) % 9 + 9) % 9
-      setSelectedSlot(newSlot)
+      if (newSlot !== bot.quickBarSlot) bot.setQuickBarSlot(newSlot)
     }, {
       passive: false,
-      signal: controller.signal
+      signal: controller.signal,
     })
 
     document.addEventListener('keydown', (e) => {
       if (!isInRealGameSession()) return
       const numPressed = +((/Digit(\d)/.exec(e.code))?.[1] ?? -1)
       if (numPressed < 1 || numPressed > 9) return
-      setSelectedSlot(numPressed - 1)
+      const newSlot = numPressed - 1
+      if (newSlot !== bot.quickBarSlot) bot.setQuickBarSlot(newSlot)
     }, {
       passive: false,
-      signal: controller.signal
+      signal: controller.signal,
     })
 
     let touchStart = 0
     document.addEventListener('touchstart', (e) => {
-      if ((e.target as HTMLElement).closest('.hotbar')) {
-        touchStart = Date.now()
-      } else {
-        touchStart = 0
-      }
-    })
+      touchStart = (e.target as HTMLElement).closest('.hotbar') ? Date.now() : 0
+    }, { signal: controller.signal })
     document.addEventListener('touchend', (e) => {
       if (touchStart && (e.target as HTMLElement).closest('.hotbar') && Date.now() - touchStart > 700) {
         triggerCommand('general.dropStack', true)
         triggerCommand('general.dropStack', false)
       }
       touchStart = 0
-    })
+    }, { signal: controller.signal })
+
+    const refresh = () => {
+      clearInventoryCaches()
+      setTextureVersion(v => v + 1)
+    }
+    appViewer.resourcesManager.on('assetsTexturesUpdated', refresh)
+    appViewer.resourcesManager.on('assetsInventoryReady', refresh)
 
     return () => {
-      inv.destroy()
       controller.abort()
-      appViewer.resourcesManager.off('assetsTexturesUpdated', upHotbarItems)
+      bot.off('heldItemChanged' as any, heldItemChanged)
+      appViewer.resourcesManager.off('assetsTexturesUpdated', refresh)
+      appViewer.resourcesManager.off('assetsInventoryReady', refresh)
     }
   }, [])
 
@@ -217,13 +176,26 @@ const HotbarInner = () => {
         }}>
         <div
           className='hotbar'
-          ref={container}
           style={{
             position: 'absolute',
             pointerEvents: 'none',
-            bottom: 'var(--hud-bottom-raw)'
+            bottom: 'var(--hud-bottom-raw)',
           }}
-        />
+        >
+          <TextureProvider config={textureConfig}>
+            <ScaleProvider scale={appScale}>
+              <InventoryProvider connector={connector}>
+                <InventoryWindow
+                  type="hotbar"
+                  properties={{
+                    showOffhand: supportsOffhand ? 1 : 0,
+                    container: isMobile ? 1 : 0,
+                  }}
+                />
+              </InventoryProvider>
+            </ScaleProvider>
+          </TextureProvider>
+        </div>
       </div>
     </Portal>
   </SharedHudVars>
@@ -232,13 +204,11 @@ const HotbarInner = () => {
 export default () => {
   const [gameMode, setGameMode] = useState(bot.game?.gameMode ?? 'creative')
   useEffect(() => {
-    bot.on('game', () => {
-      setGameMode(bot.game.gameMode)
-    })
+    const onGame = () => setGameMode(bot.game.gameMode)
+    bot.on('game', onGame)
+    return () => { bot.off('game', onGame) }
   }, [])
-
-  // return gameMode === 'spectator' ? null : <HotbarInner />
-  return null // TODO!
+  return gameMode === 'spectator' ? null : <HotbarInner />
 }
 
 const Portal = ({ children, to = document.body }) => {
